rename: 카프카 관련 파일 구조 변경 #94
